# Faster IQ1_BN Metal implementation